FEAT: add policy checks for update, edit methods in UserController

+++ app/Http/Controllers/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\UpdateUserRequest;
 use App\Models\User;
 use Auth;
 use Storage;
@@ -14,8 +15,8 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
-        $ideas=$user->ideas()->paginate(5);
-        return view('users.show',compact('user','ideas'));
+        $ideas = $user->ideas()->paginate(5);
+        return view('users.show', compact('user', 'ideas'));
     }
 
     /**
@@ -23,33 +24,30 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
-        $ideas=$user->ideas()->paginate(5);
-        return view('users.edit',compact('user','ideas'));
+        $this->authorize('update', $user);
+        $ideas = $user->ideas()->paginate(5);
+        return view('users.edit', compact('user', 'ideas'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(User $user)
+    public function update(UpdateUserRequest $request, User $user)
     {
-       $validated=request()->validate(
-        [
-            'name'=>'nullable|min:3|max:40',
-            'bio'=>'nullable|min:3|max:255',
-            'image'=>'image'
-        ]
-    );
-    if(request()->has('image'))
-    {
-        $imagepath=request()->file('image')->store('profile','public');
-        $validated['image']=$imagepath;
-        Storage::disk('public')->delete($user->image??'');
-    }
+        $this->authorize('update', $user);   
+        $validated = $request->validated();
+
+        if (request()->has('image')) {
+            $imagepath = request()->file('image')->store('profile', 'public');
+            $validated['image'] = $imagepath;
+            Storage::disk('public')->delete($user->image ?? '');
+        }
         $user->update($validated);
         return redirect()->route('profile');
     }
 
-    public function profile(){
+    public function profile()
+    {
         return $this->show(Auth::user());
     }
 }
